+++ examples/test/test_all_examples.py
@@ -1276,14 +1276,20 @@ def test_pyomo_heuristic_dispatch_example(subtests, temp_copy_of_example):
         )
         assert electricity_unmet_demand == pytest.approx(711.1997294551337, rel=1e-6)
 
-    # check that error is raised when incorrect tech_name is given
-    with subtests.test("Check incorrect tech_name error"):
-        expected_error = (
-            r"tech_name in control_parameters \(wrong_tech_name\) must match "
-            r"the top-level name of the tech group \(battery\)"
-        )
-        with pytest.raises(ValueError, match=expected_error):
-            H2IntegrateModel(example_folder / "pyomo_heuristic_dispatch_error_for_testing.yaml")
+    # Check that incorrect and no tech name provided will be replaced and validate
+    model_config = load_yaml(example_folder / "pyomo_heuristic_dispatch.yaml")
+    tech = load_yaml(example_folder / "tech_config.yaml")
+    with subtests.test("Ensure no-tech name validates"):
+        tech["technologies"]["battery"]["model_inputs"]["control_parameters"] = None
+        model_config["technology_config"] = tech
+        model = H2IntegrateModel(model_config)
+
+    with subtests.test("Ensure incorrect name is corrected"):
+        tech["technologies"]["battery"]["model_inputs"]["control_parameters"] = {
+            "tech_name": "goose"
+        }
+        model_config["technology_config"] = tech
+        model = H2IntegrateModel(model_config)
 
 
 @pytest.mark.integration