+++ h2integrate/core/h2integrate_model.py
@@ -188,6 +188,21 @@ class H2IntegrateModel:
             )
         )
 
+        for name, vals in self.technology_config["technologies"].items():
+            if "control_parameters" in vals["model_inputs"]:
+                val = self.technology_config["technologies"][name]["model_inputs"][
+                    "control_parameters"
+                ]
+                updated = {"tech_name": name}
+                if val is not None:
+                    self.technology_config["technologies"][name]["model_inputs"][
+                        "control_parameters"
+                    ].update(updated)
+                else:
+                    self.technology_config["technologies"][name]["model_inputs"][
+                        "control_parameters"
+                    ] = updated
+
     def create_custom_models(self, model_config, config_parent_path, model_types, prefix=""):
         """This method loads custom models from the specified directory and adds them to the
         supported models dictionary.