+++ CHANGELOG.md
@@ -20,6 +20,7 @@
   - Partial testing suite refactor to parameterize many of the common fixtures and test routines.
   - `unittest` style tests are refactored to be `pytest` style tests for test consistency.
 - Added a pre-commit hook for `yamlfix` to auto-format YAML files and `yamlfix`'d all YAML files for consistent formatting
+- Insert model names for technologies with control strategies to simplify Pyomo workflows.
 
 ## 0.6 [February 10, 2026]